# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/invoices/{invoiceId}/payments": {
      "post": {
        "operationId": "MoneyController_recordPayment",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      },
      "get": {
        "operationId": "MoneyController_listPayments",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/invoices/{invoiceId}/verify": {
      "post": {
        "operationId": "MoneyController_verifyPayment",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/invoices": {
      "get": {
        "operationId": "MoneyController_listInvoices",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "customer_due",
                "paid",
                "payment_verified",
                "in_payout",
                "bill_cleared"
              ],
              "type": "string"
            }
          },
          {
            "name": "cashPending",
            "required": false,
            "in": "query",
            "description": "Only invoices with a cash payment nobody has verified yet — the queue the\nPayments page works through.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "id of the last row from the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "The boat's invoices (owner console). Read-only view of the money list.",
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/refunds": {
      "get": {
        "operationId": "MoneyController_listRefunds",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/invoices/{invoiceId}/refunds": {
      "post": {
        "operationId": "MoneyController_requestRefund",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/refunds/{refundId}/verify": {
      "post": {
        "operationId": "MoneyController_verifyRefund",
        "parameters": [
          {
            "name": "refundId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/refunds/{refundId}/complete": {
      "post": {
        "operationId": "MoneyController_completeRefund",
        "parameters": [
          {
            "name": "refundId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/due-payments": {
      "get": {
        "operationId": "MoneyController_duePayments",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/payout-batches": {
      "get": {
        "operationId": "MoneyController_listPayoutBatches",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Payout history, owner-readable. Preparing/approving/paying stay\nplatform-only above — an owner sees what they were paid, they don't move it.",
        "tags": [
          "Money"
        ]
      },
      "post": {
        "operationId": "MoneyController_prepareBatch",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/payout-batches/{batchId}/approve": {
      "post": {
        "operationId": "MoneyController_approveBatch",
        "parameters": [
          {
            "name": "batchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/payout-batches/{batchId}/pay": {
      "post": {
        "operationId": "MoneyController_payBatch",
        "parameters": [
          {
            "name": "batchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/coupons": {
      "get": {
        "operationId": "MoneyController_listCoupons",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      },
      "post": {
        "operationId": "MoneyController_createCoupon",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/cancellation-policies": {
      "get": {
        "operationId": "MoneyController_listPolicies",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      },
      "post": {
        "operationId": "MoneyController_createPolicy",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/distributions": {
      "get": {
        "operationId": "MoneyController_listDistributions",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      },
      "post": {
        "operationId": "MoneyController_recordDistribution",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordDistributionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/distributions/suggest": {
      "get": {
        "operationId": "MoneyController_suggestSplit",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/billing-status": {
      "get": {
        "operationId": "MoneyController_billingStatus",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/my-subscription-invoices": {
      "get": {
        "operationId": "MoneyController_ownerSubscriptionInvoices",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Owner-visible list of their subscription invoices (to see what to pay).\n Accessible even when locked.",
        "tags": [
          "Money"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/subscription-invoices": {
      "get": {
        "operationId": "MoneyController_listSubscriptionInvoices",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      },
      "post": {
        "operationId": "MoneyController_issueSubscription",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
    "/api/subscription-invoices/{invoiceId}/pay": {
      "post": {
        "operationId": "MoneyController_paySubscription",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Money"
        ]
      }
    },
...
  },
  "info": {
    "title": "Houseboat API",
    "description": "Booking SaaS backend",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hb_access"
      }
    },
    "schemas": {
...
      "RecordPaymentDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "method": {
            "type": "string",
            "enum": [
              "gateway",
              "cash"
            ]
          },
          "gatewayToken": {
            "type": "string",
            "maxLength": 64
          },
          "receivedBy": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "amount",
          "method"
        ]
      },
      "RefundRequestDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          },
          "bankDetails": {
            "type": "object",
            "description": "Encrypted at rest (AES-256-GCM) before storage — see refunds.service.ts.\nShape is provider-dependent, so it stays an object rather than a fixed DTO;\n@IsObject at least rejects scalars and arrays."
          }
        },
        "required": [
          "amount"
        ]
      },
      "CreateCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64
          },
          "kind": {
            "type": "string",
            "enum": [
              "percent",
              "flat",
              "referral"
            ]
          },
          "value": {
            "type": "number",
            "minimum": 0
          },
          "validFrom": {
            "type": "string"
          },
          "validTo": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "kind",
          "value"
        ]
      },
      "CreatePolicyDto": {
        "type": "object",
        "properties": {
          "policyTemplate": {
            "type": "string",
            "enum": [
              "flexible",
              "moderate",
              "strict",
              "non_refundable",
              "custom"
            ]
          },
          "depositPct": {
            "type": "number",
            "minimum": 0
          },
          "tiers": {
            "type": "object",
            "description": "Refund tier table; shape varies by template."
          }
        },
        "required": [
          "policyTemplate"
        ]
      },
      "RecordDistributionDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "maxLength": 64
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "membershipId",
          "amount"
        ]
      },
      "IssueSubscriptionDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "description": "YYYY-MM",
            "maxLength": 64
          }
        },
        "required": [
          "period"
        ]
      },
...
    }
  }
}